pyproject.toml
[tool.poetry]
name = "42musaev"
version = "0.1.0"
description = ""
authors = ["Askhab Musaev <42musaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
